requirements.txt
# AWS Lambda + Playwright + Chromium 완전한 패키지 목록
playwright==1.40.0
PyYAML>=6.0
holidays>=0.34
pydantic>=2.0.0
requests>=2.31.0
boto3>=1.26.0
typing-extensions>=4.5.0
python-dateutil>=2.8.0

# 웹 자동화 관련
beautifulsoup4>=4.12.0
lxml>=4.9.0
selenium>=4.10.0

# 유틸리티
python-dotenv>=1.0.0
loguru>=0.7.0

# 날짜/시간 처리
pytz>=2023.3
dateparser>=1.1.0


aiohttp==3.9.1
